Move updateRecentValues into update-utilities.
